requirements.txt
torch
torchvision
transformers
gradio
pytesseract
Pillow
streamlit
